Updating the S-125 package version

+++ niord-uk-s125/src/main/java/org/niord/s125/utils/S125DatasetBuilder.java
@@ -1409,7 +1409,7 @@ public class S125DatasetBuilder {
                 .map(atonNode::getTag)
                 .map(AtonTag::getV)
                 .filter(StringUtils::isNotBlank)
-                .map(BigInteger::new)
+                .map(BigDecimal::new)
                 .orElse(null));
         member.setMMSICode(Optional.of(tagKeyPrefix+"mmsi")
                 .map(atonNode::getTag)
